+++ config/config.php
@@ -11,10 +11,10 @@
  * @copyright Frank Hoppe 2014
  */
 
-define(CACHE_AKTIV, true); // Cachestatus
-define(CACHE_TIME, 36000); // Cachezeit (36000 = 10h)
-define(CACHE_TIME_FAKTOR_VERBAND, 100); // CACHE_TIME * Faktor = gesamte Cachelebenszeit
-define(CACHE_TIME_FAKTOR_REFERENT, 20); // CACHE_TIME * Faktor = gesamte Cachelebenszeit
+//define(CACHE_AKTIV, false); // Cachestatus
+//define(CACHE_TIME, 36000); // Cachezeit (36000 = 10h)
+//define(CACHE_TIME_FAKTOR_VERBAND, 20); // CACHE_TIME * Faktor = gesamte Cachelebenszeit
+//define(CACHE_TIME_FAKTOR_REFERENT, 5); // CACHE_TIME * Faktor = gesamte Cachelebenszeit
 define(CACHE_DIR, TL_ROOT . '/system/cache/dewis/'); // Cacheverzeichnis festlegen
 
 define(KARTEISPERRE_GAESTE, true); // Anzeige von Karteikarten für nichtangemeldete Besucher gesperrt
@@ -93,3 +93,14 @@ if (TL_MODE == 'BE')
 	//print_r(get_defined_constants());
 	//echo "</pre>";
 }
+
+/**
+ * -------------------------------------------------------------------------
+ * Voreinstellungen Contao-BE System -> Einstellungen
+ * -------------------------------------------------------------------------
+ */
+
+$GLOBALS['TL_CONFIG']['dewis_cache'] = 1;
+$GLOBALS['TL_CONFIG']['dewis_cache_default'] = 4;
+$GLOBALS['TL_CONFIG']['dewis_cache_referent'] = 24;
+$GLOBALS['TL_CONFIG']['dewis_cache_verband'] = 48;